requirements.txt
# SOCPilot dependencies
# ── Static file serving ───────────────────────────────────────────────────────
aiofiles>=23.0.0           # Required by FastAPI for FileResponse / StaticFiles


# ── AI inference ──────────────────────────────────────────────────────────────
anthropic>=0.40.0          # Anthropic SDK  (AI_PROVIDER=anthropic, default)
openai>=1.30.0             # OpenAI-compatible SDK (AI_PROVIDER=openai → Ollama/LM Studio/vLLM)

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0       # Load .env in development

# ── HTTP API server ───────────────────────────────────────────────────────────
fastapi>=0.115.0           # REST + WebSocket framework
uvicorn[standard]>=0.30.0  # ASGI server (includes websockets, httptools)
pydantic>=2.0.0            # Request/response models (bundled with FastAPI)